IGNITE-18079 / make Cluster a top-level class
